requirements.txt
attrs==18.1.0
Automat==0.7.0
certifi==2018.4.16
chardet==3.0.4
constantly==15.1.0
docutils==0.14
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
lockfile==0.12.2
pkg-resources==0.0.0
python-daemon==2.1.2
requests==2.19.1
six==1.11.0
Twisted==18.4.0
urllib3==1.23
zope.interface==4.5.0
